requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
Django==2.2
django-allauth==0.42.0
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-model-utils==4.0.0
django-mptt==0.11.0
django-notifications-hq==1.6.0
django-password-reset==2.0
django-taggit==1.3.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.1
jsonfield==3.1.0
Markdown==3.2.2
mysqlclient==1.4.6
oauthlib==3.1.0
Pillow==7.1.2
Pygments==2.6.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
swapper==1.1.2.post1
urllib3==1.25.9
zipp==3.1.0
